+++ ipspot/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 """ipspot modules."""
 from .params import IPSPOT_VERSION, IPv4API
-from .functions import get_private_ipv4, get_public_ipv4
+from .functions import get_private_ipv4, get_public_ipv4, is_ipv4
 __version__ = IPSPOT_VERSION