+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `is_ipv4` function
 - Support [ident.me](https://ident.me/json)
 ### Changed
 - `get_private_ipv4` function modified